Update stow directory structure. Use directories for each config
Stow using `stow *` or `stow */` to only symlink directories. The latter
is useful when dotfiles directory has a README.md or other files that
should not be symlinked.

Using `stow *` stows packages correctly by stowing the contents of each
package directory in the dotfiles repo directly to the home directory,
while using `stow .` symlinks the dotfiles repo as is which would be
incorrect when using the conventional stow directory structure