Add support for signing commits with SSH key defined in `local.config`
The path to the SSH key is defined in a local.config file which is
ignored in `.gitignore`. The path to `local.config` is included in
`git/config`, ensuring that the local key path can be referenced on any
system with it's own `.config/git/local.config` from the git tracked
`.config/git/config` file.